backend/requirements.txt
# System Dependencies (must be first)
setuptools>=68.0.0
wheel>=0.40.0

# Core FastAPI and Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# AI and Blockchain (with compatible versions)
google-generativeai==0.3.2
web3==6.11.3
py-solc-x==2.0.4
eth-utils>=2.0.0
eth-typing>=3.0.0

# Monitoring and Metrics
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0

# HTTP Client
httpx==0.25.2
requests==2.32.3

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
python-multipart==0.0.6

# Email & Communication
fastapi-mail==1.4.1
jinja2==3.1.2

# Rate Limiting & Security
slowapi==0.1.9
python-limits==3.5.0

# OAuth Integration
authlib==1.2.1
httpx-oauth==0.10.2

# JWT & Session Management
pyjwt==2.8.0
redis==5.0.1

# Validation & Serialization
email-validator==2.1.0

# Utilities
PyYAML==6.0.2

# Security and Compliance
cryptography==42.0.5
bcrypt==4.3.0
pyopenssl==24.1.0
python-multipart==0.0.6

# Redis for caching and rate limiting
redis==5.0.1
hiredis==2.3.2